+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [0.2.3] - 2021-11-18
 
+- Fix samples not working
+
 ## [0.2.1] - 2021-09-23
 
 - Test and document sql metrics