+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## [0.3.0] - 2021-12-22
+
 
 - Provides the ability to get the scan results in Dataframes. ([#99](https://github.com/sodadata/soda-spark/pull/99))
   - measurements